pyproject.toml
[project]
name = "chemistry-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.120.4",
    "fastembed>=0.7.3",
    "google-genai>=1.47.0",
    "gradio>=5.49.1",
    "huggingface-hub>=0.36.0",
    "langchain>=1.0.3",
    "langchain-google-genai>=3.0.0",
    "langgraph>=1.0.2",
    "numpy>=2.3.4",
    "pillow>=11.3.0",
    "piper-tts>=1.3.0",
    "pydantic>=2.11.10",
    "pymongo>=4.15.3",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.15.1",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.51.0",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.3",
]
